Removes tests that weren't working
